cloud-function-recents-fetcher/requirements.txt
requests
beautifulsoup4
pillow
google-cloud-storage
google-cloud-datastore
grpcio